# openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MeisterProPR.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/admin/users/{id}/clients/{clientId}": {
      "delete": {
        "tags": [
          "AdminUsers"
        ],
        "summary": "Removes a client role assignment from a user.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/clients/{clientId}/ai-connections/discover-models": {
      "post": {
        "tags": ["ClientAiConnections"],
        "summary": "Probes the given endpoint URL and returns available model deployment names.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DiscoverModelsRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of model deployment names (may be empty).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DiscoverModelsResponse" }
              }
            }
          },
          "400": { "description": "Validation failure." },
          "401": { "description": "Missing or invalid credentials." },
          "403": { "description": "Caller does not own this client." }
        }
      }
    },
    "/clients/{clientId}/ai-connections": {
      "get": {
        "tags": ["ClientAiConnections"],
        "summary": "Lists all AI connections for the specified client.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of AI connections (API keys are never returned).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/AiConnectionDto" }
                }
              }
            }
          },
          "401": { "description": "Missing or invalid credentials." },
          "403": { "description": "Caller does not own this client." }
        }
      },
      "post": {
        "tags": ["ClientAiConnections"],
        "summary": "Creates a new AI connection for the specified client.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateAiConnectionRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "AI connection created.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AiConnectionDto" }
              }
            }
          },
          "400": { "description": "Validation failure." },
          "401": { "description": "Missing or invalid credentials." },
          "403": { "description": "Caller does not own this client." }
        }
      }
    },
    "/clients/{clientId}/ai-connections/{connectionId}": {
      "patch": {
        "tags": ["ClientAiConnections"],
        "summary": "Updates non-null fields of an AI connection.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "connectionId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateAiConnectionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI connection updated.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AiConnectionDto" }
              }
            }
          },
          "400": { "description": "Validation failure." },
          "401": { "description": "Missing or invalid credentials." },
          "403": { "description": "Caller does not own this client." },
          "404": { "description": "AI connection not found." }
        }
      },
      "delete": {
        "tags": ["ClientAiConnections"],
        "summary": "Deletes an AI connection.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "connectionId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "204": { "description": "AI connection deleted." },
          "401": { "description": "Missing or invalid credentials." },
          "403": { "description": "Caller does not own this client." },
          "404": { "description": "AI connection not found." }
        }
      }
    },
    "/clients/{clientId}/ai-connections/{connectionId}/activate": {
      "post": {
        "tags": ["ClientAiConnections"],
        "summary": "Activates the specified AI connection with the given model deployment.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "connectionId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ActivateAiConnectionRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI connection activated.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AiConnectionDto" }
              }
            }
          },
          "400": { "description": "Model not in the connection's model list." },
          "401": { "description": "Missing or invalid credentials." },
          "403": { "description": "Caller does not own this client." },
          "404": { "description": "AI connection not found." }
        }
      }
    },
    "/clients/{clientId}/ai-connections/{connectionId}/deactivate": {
      "post": {
        "tags": ["ClientAiConnections"],
        "summary": "Deactivates the specified AI connection.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          },
          {
            "name": "connectionId",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "AI connection deactivated.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AiConnectionDto" }
              }
            }
          },
          "401": { "description": "Missing or invalid credentials." },
          "403": { "description": "Caller does not own this client." },
          "404": { "description": "AI connection not found." }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AiConnectionModelCategory": {
        "enum": [
          "lowEffort",
          "mediumEffort",
          "highEffort"
        ],
        "type": "string",
        "description": "Model category tag for tier-based routing. lowEffort for small diffs, mediumEffort for mid-size diffs, highEffort for large diffs."
      },
...
      "AiConnectionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "nullable": true
          },
          "endpointUrl": {
            "type": "string",
            "nullable": true
          },
          "models": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "activeModel": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "modelCategory": {
            "$ref": "#/components/schemas/AiConnectionModelCategory",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Data transfer object for an AI connection. API key is never returned."
      },
      "CreateAiConnectionRequest": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true,
            "description": "Human-readable name. Required, 1-200 chars."
          },
          "endpointUrl": {
            "type": "string",
            "nullable": true,
            "description": "Azure OpenAI or AI Foundry endpoint URL. Required, immutable after creation."
          },
          "models": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "description": "Available model deployment names. Required, non-empty."
          },
          "apiKey": {
            "type": "string",
            "nullable": true,
            "description": "Optional API key. Omit to use DefaultAzureCredential."
          },
          "modelCategory": {
            "$ref": "#/components/schemas/AiConnectionModelCategory",
            "nullable": true,
            "description": "Optional model category for tier-based routing. Null means default connection."
          }
        },
        "additionalProperties": false,
        "description": "Request body for creating a new AI connection."
      },
      "UpdateAiConnectionRequest": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "nullable": true,
            "description": "New display name. Omit to leave unchanged."
          },
          "models": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "description": "New model list. Omit to leave unchanged."
          },
          "apiKey": {
            "type": "string",
            "nullable": true,
            "description": "New API key. Null clears the stored key (reverts to DefaultAzureCredential). Omit to leave unchanged."
          }
        },
        "additionalProperties": false,
        "description": "Request body for updating an AI connection. All fields are optional."
      },
      "ActivateAiConnectionRequest": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string",
            "nullable": true,
            "description": "Model deployment name to activate. Must be in the connection's models list."
          }
        },
        "additionalProperties": false,
        "description": "Request body for activating an AI connection."
      },
      "DiscoverModelsRequest": {
        "type": "object",
        "properties": {
          "endpointUrl": { "type": "string" },
          "apiKey": { "type": "string" }
        },
        "additionalProperties": false,
        "description": "Request body for probing an AI endpoint to discover available model deployments."
      },
      "DiscoverModelsResponse": {
        "type": "object",
        "properties": {
          "models": {
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "additionalProperties": false,
        "description": "Response from the discover-models probe."
      },
...
    }
  },
...
}